+++ lib/ui/login_page.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import '../main.dart';
 import '../state/app_state.dart';
 import 'dart:math'; // For random pet picture
+import '../main_nav.dart';
 
 class LoginPage extends StatefulWidget {
   const LoginPage({super.key});
